test pull

+++ 2230912_2130331_Lab5Partie2/2230912_2130331_Lab5Partie2/Models/Cours.cs
@@ -17,7 +17,7 @@
         /// <summary>
         /// Durée du cours
         /// </summary>
-        public double DureeCours { get; set; }  
+        public int DureeCours { get; set; }  
 
 
         //Constructeur vide requis pour la désérialisation
